+++ decomp/ST.exe/functions/00604A90/decomp.c
@@ -79,7 +79,7 @@ undefined4 * __thiscall STExplosionC::SaveObj(STExplosionC *this,uint *param_1)
       local_3c[0] = (uint *)FUN_006b0020((uint *)this->field_0215,(int *)&local_20);
       uVar9 = *param_1 + local_20 + 4;
       *param_1 = uVar9;
-      local_48 = (undefined4 *)Library::DKW::LIB::FUN_006acf50(local_48,uVar9);
+      local_48 = (undefined4 *)Library::DKW::LIB::FUN_006acf50((int)local_48,uVar9);
       *(uint *)((*param_1 - local_20) + -4 + (int)local_48) = local_20;
       puVar5 = local_3c[0];
       puVar11 = (uint *)((*param_1 - local_20) + (int)local_48);
@@ -123,7 +123,7 @@ undefined4 * __thiscall STExplosionC::SaveObj(STExplosionC *this,uint *param_1)
       }
       uVar9 = *param_1;
       *param_1 = uVar9 + iVar3;
-      local_48 = (undefined4 *)Library::DKW::LIB::FUN_006acf50(local_48,uVar9 + iVar3);
+      local_48 = (undefined4 *)Library::DKW::LIB::FUN_006acf50((int)local_48,uVar9 + iVar3);
       local_44 = (uint *)((*param_1 - iVar3) + (int)local_48);
       local_30 = 0;
       if (0 < (int)this->field_0269) {
